+++ src/game/tests.rs
@@ -400,6 +400,7 @@ mod vintage_cube_faerie_mastermind;
 mod vintage_cube_fanatic_of_rhonas;
 mod vintage_cube_fatal_push;
 mod vintage_cube_fear_of_missing_out;
+mod vintage_cube_field_of_the_dead;
 mod vintage_cube_fiery_confluence;
 mod vintage_cube_figure_of_destiny;
 mod vintage_cube_fire_covenant;